Improve GMusic album name parsing.
This commit fixes a possible crash when handling GMusic albums names
and includes a small improvement to the overall parsing logic.

Normally, GMusic album names come in the form of "artist - album",
which means the album name can be found by splitting on the dash.
However, some entries do not have an artist for some reason (perhaps
old data or some weird internal quirk of the Google APIs). The code is
now changed to suppress the exception resulting from being unable to
parse these names.

The parser will now also pull the last entry in the split array rather
than the second, because there might be some weird artist out there
that has a " - " as part of their artist name. It is most likely that
the album would appear last in the split array, should multiple " - "
tokens be encountered.